Feature cxx (#221)
* Restructured code and introduced cross-platform build to fix #184 
* Separated C++ and Python workflows
* Fixed minor module structure problems

+++ norse/torch/module/test/test_lif_correlation.py
@@ -117,7 +117,7 @@ def test_lif_correlation_training():
         m, _ = torch.max(voltages, dim=0)
 
         log_p_y = log_softmax_fn(m)
-        loss_val = loss_fn(log_p_y, y_data)
+        loss_val = loss_fn(log_p_y, y_data.long())
 
         loss_val.backward()
         optimizer.step()
